requirements.txt
Flask==3.0.0
pycryptodome==3.18.0
Pillow==10.0.1